+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-1-10
+
+### Removed
+
+- csstype removed in favour of React
+
+### Added
+
+- React as preferred component creator
+- Pre-made components: Div, Heading, Image, List, Paragraph
+
 ## [1.0.2] - 2024-1-10
 
 ### Removed